fix: extract wishlist server action to dedicated module
Inline server actions that close over component props can cause RSC
stream serialization issues in Next.js 16 with cacheComponents:true.
Move toggleWishlist into wishlist-actions.ts and bind productId via
.bind() — the recommended Next.js pattern for bound server actions.

Also add await connection() to WishlistSection so PPR correctly treats
it as a dynamic component, preventing stale static pre-rendering.
