+++ package.json
@@ -88,7 +88,12 @@
 			"@babel/runtime": ">=7.26.10",
 			"@babel/helpers": ">=7.26.10",
 			"brace-expansion@>=2": ">=2.0.2"
-		}
+		},
+		"onlyBuiltDependencies": [
+			"esbuild",
+			"sharp",
+			"unrs-resolver"
+		]
 	},
 	"packageManager": "pnpm@10.28.1",
 	"engines": {
