+++ src/ui/components/pdp/wishlist-section.tsx
@@ -1,7 +1,9 @@
+import { connection } from "next/server";
 import { executeAuthenticatedGraphQL } from "@/lib/graphql";
-import { WishlistFetchDocument, WishlistUpdateDocument } from "@/gql/graphql";
+import { WishlistFetchDocument } from "@/gql/graphql";
 import { parseWishlist } from "./wishlist-utils";
 import { AddToWishlist } from "./add-to-wishlist";
+import { updateWishlist } from "./wishlist-actions";
 
 interface WishlistSectionProps {
 	productId: string;
@@ -9,6 +11,12 @@ interface WishlistSectionProps {
 }
 
 export async function WishlistSection({ productId, channel }: WishlistSectionProps) {
+	// Opt into per-request dynamic rendering — cookies() access is buried inside
+	// executeAuthenticatedGraphQL and Next.js static analysis won't detect it.
+	// Without this, cacheComponents:true may try to pre-render this component
+	// statically, which scrambles the Suspense lazy slot on the client.
+	await connection();
+
 	const result = await executeAuthenticatedGraphQL(WishlistFetchDocument, {
 		cache: "no-cache",
 	});
@@ -17,39 +25,12 @@ export async function WishlistSection({ productId, channel }: WishlistSectionPro
 	const wishlist = parseWishlist(user?.metafield);
 	const isWishlisted = wishlist.includes(productId);
 
-	async function toggleWishlist() {
-		"use server";
-
-		const current = await executeAuthenticatedGraphQL(WishlistFetchDocument, {
-			cache: "no-cache",
-		});
-
-		if (!current.ok || !current.data.me) return;
-
-		const currentList = parseWishlist(current.data.me.metafield);
-		const updated = currentList.includes(productId)
-			? currentList.filter((id) => id !== productId)
-			: [...currentList, productId];
-
-		const updateResult = await executeAuthenticatedGraphQL(WishlistUpdateDocument, {
-			variables: {
-				id: current.data.me.id,
-				input: [{ key: "wishlist", value: JSON.stringify(updated) }],
-			},
-			cache: "no-cache",
-		});
-
-		if (!updateResult.ok) {
-			console.error("Wishlist update failed:", updateResult.error.message);
-		}
-	}
-
 	return (
 		<AddToWishlist
 			isWishlisted={isWishlisted}
 			isLoggedIn={!!user}
 			channel={channel}
-			action={toggleWishlist}
+			action={updateWishlist.bind(null, productId)}
 		/>
 	);
 }
